+++ src/index.ts
@@ -1,36 +1,10 @@
 import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
-
-import { TenableIntegrationConfig } from './config';
-import invocationValidator from './invocationValidator';
-import { accountStep } from './steps/account';
-import { containerSteps } from './steps/containers';
-import { scanSteps } from './steps/vulnerabilities';
-import { userStep } from './steps/access';
-import { serviceSteps } from './steps/service';
-
-export const invocationConfig: IntegrationInvocationConfig<TenableIntegrationConfig> =
+import { IntegrationConfig, instanceConfigFields } from './config';
+import validateInvocation from './invocationValidator';
+import { integrationSteps } from './steps';
+export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> =
   {
-    instanceConfigFields: {
-      accessKey: {
-        type: 'string',
-      },
-      secretKey: {
-        type: 'string',
-        mask: true,
-      },
-      assetApiTimeoutInMinutes: {
-        type: 'string',
-      },
-      vulnerabilityApiTimeoutInMinutes: {
-        type: 'string',
-      },
-    },
-    validateInvocation: invocationValidator,
-    integrationSteps: [
-      accountStep,
-      ...serviceSteps,
-      ...containerSteps,
-      ...scanSteps,
-      userStep,
-    ],
+    instanceConfigFields,
+    validateInvocation,
+    integrationSteps,
   };