+++ src/steps/vulnerabilities/index.ts
@@ -5,13 +5,13 @@ import {
   RelationshipClass,
   Step,
 } from '@jupiterone/integration-sdk-core';
-import { TenableIntegrationConfig } from '../../config';
+import { IntegrationConfig } from '../../config';
 import {
   Entities,
   MappedRelationships,
   Relationships,
   StepIds,
-} from '../../constants';
+} from '../constants';
 import TenableClient from '../../tenable/TenableClient';
 import { AssetExport, VulnerabilityExport } from '../../tenable/client';
 import {
@@ -24,11 +24,12 @@ import {
   createRelationshipFromTargetEntity,
   createRelationshipToTargetEntity,
 } from '../../utils/targetEntities';
-import { getAccount } from '../../initializeContext';
+import { getAccount } from '../account/util';
 import { createAccountEntity } from '../account/converters';
+import { buildFilters } from './filters';
 
 export async function fetchAssets(
-  context: IntegrationStepExecutionContext<TenableIntegrationConfig>,
+  context: IntegrationStepExecutionContext<IntegrationConfig>,
 ): Promise<void> {
   const { jobState, logger, instance } = context;
   const accountEntity = createAccountEntity(getAccount(context));
@@ -42,16 +43,18 @@ export async function fetchAssets(
 
   logger.info({ assetApiTimeoutInMinutes }, 'Attempting to fetch assets...');
 
+  let duplicateKeysEncountered = 0;
   await provider.iterateAssets(
     async (asset) => {
       const assetEntity = createAssetEntity(asset, logger);
       if (await jobState.hasKey(assetEntity._key)) {
-        logger.warn(
+        logger.debug(
           {
             _key: assetEntity._key,
           },
-          'Warning: duplicate asset _key encountered',
+          'Debug: duplicate asset _key encountered',
         );
+        duplicateKeysEncountered += 1;
         return;
       }
       await jobState.addEntity(assetEntity);
@@ -78,10 +81,17 @@ export async function fetchAssets(
       timeoutInMinutes: assetApiTimeoutInMinutes,
     },
   );
+
+  if (duplicateKeysEncountered > 0) {
+    logger.info(
+      { duplicateKeysEncountered },
+      `Found duplicate keys for "tenable_asset" entity`,
+    );
+  }
 }
 
 export async function fetchVulnerabilities(
-  context: IntegrationStepExecutionContext<TenableIntegrationConfig>,
+  context: IntegrationStepExecutionContext<IntegrationConfig>,
 ): Promise<void> {
   const { jobState, logger, instance } = context;
   const { vulnerabilityApiTimeoutInMinutes, accessKey, secretKey } =
@@ -97,29 +107,42 @@ export async function fetchVulnerabilities(
     'Attempting to fetch vulnerabilities...',
   );
 
+  let duplicateKeysEncountered = 0;
   await provider.iterateVulnerabilities(
     async (vuln) => {
       // TODO add `targets` property from the asset.
       const vulnerabilityEntity = createVulnerabilityEntity(vuln, [], logger);
       if (await jobState.hasKey(vulnerabilityEntity._key)) {
-        logger.warn(
+        logger.debug(
           {
             _key: vulnerabilityEntity._key,
           },
-          'Warning: duplicate tenable_vulnerability_finding _key encountered',
+          'Debug: duplicate tenable_vulnerability_finding _key encountered',
         );
+        duplicateKeysEncountered += 1;
         return;
       }
       await jobState.addEntity(vulnerabilityEntity);
     },
     {
       timeoutInMinutes: vulnerabilityApiTimeoutInMinutes,
+      exportVulnerabilitiesOptions: {
+        num_assets: 50,
+        filters: buildFilters(instance.config),
+      },
     },
   );
+
+  if (duplicateKeysEncountered > 0) {
+    logger.info(
+      { duplicateKeysEncountered },
+      `Found duplicate keys for "tenable_vulnerability_finding" entity`,
+    );
+  }
 }
 
 export async function buildAssetVulnerabilityRelationships(
-  context: IntegrationStepExecutionContext<TenableIntegrationConfig>,
+  context: IntegrationStepExecutionContext<IntegrationConfig>,
 ): Promise<void> {
   const { jobState, logger } = context;
 
@@ -139,7 +162,7 @@ export async function buildAssetVulnerabilityRelationships(
 
       const assetEntity = await jobState.findEntity(vulnRawData.asset.uuid);
       if (!assetEntity) {
-        logger.warn(
+        logger.debug(
           {
             'vuln._key': vulnEntity._key,
             'asset.uuid': vulnRawData.asset.uuid,
@@ -184,10 +207,11 @@ export async function buildAssetVulnerabilityRelationships(
 }
 
 export async function buildVulnerabilityCveRelationships(
-  context: IntegrationStepExecutionContext<TenableIntegrationConfig>,
+  context: IntegrationStepExecutionContext<IntegrationConfig>,
 ): Promise<void> {
   const { jobState, logger } = context;
 
+  let duplicateKeysEncountered = 0;
   await jobState.iterateEntities(
     { _type: Entities.VULNERABILITY._type },
     async (vulnEntity) => {
@@ -209,22 +233,30 @@ export async function buildVulnerabilityCveRelationships(
           to: targetCveEntity,
         });
         if (await jobState.hasKey(vulnCveMappedRelationship._key)) {
-          logger.warn(
+          logger.debug(
             {
               _key: vulnCveMappedRelationship._key,
             },
-            'Warning: duplicate tenable_vulnerability_finding_is_cve _key encountered',
+            'Debug: duplicate tenable_vulnerability_finding_is_cve _key encountered',
           );
+          duplicateKeysEncountered += 1;
           break;
         }
         await jobState.addRelationship(vulnCveMappedRelationship);
       }
     },
   );
+
+  if (duplicateKeysEncountered > 0) {
+    logger.info(
+      { duplicateKeysEncountered },
+      `Found duplicate keys for "tenable_vulnerability_finding_is_cve" relationship`,
+    );
+  }
 }
 
 export const scanSteps: Step<
-  IntegrationStepExecutionContext<TenableIntegrationConfig>
+  IntegrationStepExecutionContext<IntegrationConfig>
 >[] = [
   {
     id: StepIds.ASSETS,